// src/pages/user/regsiter/regsiter.interface.ts
export interface IRegsiterInterface {
    Name: string; // 企业名称
    ReffererCode: string; // 企业代码
    Addr: string; // 企业地址
    Tel: number; // 企业电话
    LegalPerson: string; // 企业联系人
    LoginId: string; // 企业登录名
    RecommendCode: string; // 和上面loginid一样，服务端说的
}